Functions for testing general functionality of PDB/PDBx models have been added.
In particular, the process_algorithm_data function was tested. However,
there were issues with testing specific functions related to b-factor parsing.
The problem is that such a test requires invoking the entire module, which
makes the function heavily dependent on other components. This should not be
the case; it is clearly a weak spot, and it would be good to perform a
refactoring. In the future, I plan to try rewriting the internal data-holding
classes using CPython. Therefore, for now, I will leave the parsing function as
it is and postpone its refactoring.

Regarding the alignment module, all functions there are so tightly coupled that
I believe it would be best to rewrite the entire module from scratch.